# internal/infrastructure/repository/subscription-repository/sqlcgen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package sqlcgen

import (
	"database/sql"
	"time"
)

type Subscription struct {
	ID          string       `json:"id"`
	ServiceName string       `json:"service_name"`
	Price       int64        `json:"price"`
	UserID      string       `json:"user_id"`
	StartDate   time.Time    `json:"start_date"`
	EndDate     sql.NullTime `json:"end_date"`
	CreatedAt   time.Time    `json:"created_at"`
	UpdatedAt   time.Time    `json:"updated_at"`
}